backend/requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
pypdf
docx2txt
fastapi
uvicorn
python-dotenv
pinecone 
langchain-pinecone 
langchain-groq
langchain-tavily
requests 
uuid 
gunicorn
pdfplumber
python-multipart